+++ sdmx-dl-api/src/test/java/internal/sdmxdl/web/BM25IndexTest.java
@@ -1,4 +1,4 @@
-package internal.sdmxdl.format.search;
+package internal.sdmxdl.web;
 
 import org.junit.jupiter.api.Test;