+++ sdmx-dl-format-base/src/main/java/module-info.java
@@ -1,5 +1,4 @@
 import sdmxdl.ext.Persistence;
-import sdmxdl.format.spi.SearchScoringProvider;
 
 module sdmxdl.format.base {
 
@@ -13,13 +12,7 @@ module sdmxdl.format.base {
 
     exports sdmxdl.format;
     exports sdmxdl.format.design;
-    exports sdmxdl.format.spi;
     exports sdmxdl.format.time;
 
     uses Persistence;
-    uses SearchScoringProvider;
-
-    provides SearchScoringProvider with
-            internal.sdmxdl.format.search.Bm25ScoringProvider,
-            internal.sdmxdl.format.search.TrigramScoringProvider;
 }