+++ CHANGELOG.md
@@ -10,9 +10,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - ![API] Add support of partial key [#1086](https://github.com/nbbrd/sdmx-dl/issues/1086)
+- ![API] Add hybrid search [#1251](https://github.com/nbbrd/sdmx-dl/issues/1251)
 - ![PROVIDER] Add SQL EXPLAIN ANALYZE-style execution tracing [#1244](https://github.com/nbbrd/sdmx-dl/issues/1244)
 - ![PROVIDER] Add retry property for transient network errors [#1252](https://github.com/nbbrd/sdmx-dl/issues/1252)
-- ![FORMAT] Add hybrid search [#1251](https://github.com/nbbrd/sdmx-dl/issues/1251)
 - ![SOURCE] Add source from State Statistics Service of Ukraine [#1257](https://github.com/nbbrd/sdmx-dl/issues/1257)
 
 ### Changed